Added new Files launcher.py and dboperations.py
launcher.py: This script will act as a starting point for the finished program. This script will interact with the other scripts.

dboperations.py: This script will contain all the database related methods of the program. This script will not execute by itself. It merely contains methods to be used by other scripts.